Remove unnecessary casts on zero [nfc]

+++ src/compression.c
@@ -106,7 +106,7 @@ static size_t decompress_buffer_bslz4_hdf5(char* dst,
     }
 
     if (orig_size == 0)
-        return (size_t)0;
+        return 0;
 
     const int leftover_bytes =
         orig_size % (int)(BSHUF_BLOCKED_MULT * elem_size);
@@ -188,7 +188,7 @@ static size_t decompress_buffer_lz4_hdf5(char* dst,
     }
 
     if (orig_size == 0)
-        return (size_t)0;
+        return 0;
 
     const uint64_t full_block_count = orig_size / block_size;
     const uint32_t last_block_size = orig_size % block_size;
